server/requirements.txt
flask
flask-cors
google-auth-oauthlib
google-api-python-client
openai
python-dotenv